telegram_bot/pyproject.toml
[project]
name = "telegram_bot"
version = "0.1.0"
description = "Telegram bot for game deal notifications"
authors = [
    {name = "galst1234", email = "12615933+galst1234@users.noreply.github.com"},
]
dependencies = [
    "pydantic~=2.12",
    "python-telegram-bot[job-queue]~=22.5",
    "requests~=2.32",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}


[dependency-groups]
dev = [
    "ruff~=0.14",
    "pre-commit~=4.5",
    "mypy~=1.19",
    "ipython~=9.8",
    "types-requests~=2.32",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

target-version = "py313"

preview = true


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # pylint
    "PL",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # flakes8-annotations
    "ANN",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-quotes
    "Q",
    # pep8-naming
    "N",
    # Ruff
    "RUF",
    # McCabe complexity
    "C90",
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pylint]
max-positional-args = 3

[tool.pdm]
distribution = false
